# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Agent Planner API",
    "version": "2.0.0",
    "description": "Collaborative planning API for humans and AI agents"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    },
...
  },
  "paths": {
...
    "/v1/orgs": {
      "get": {
        "summary": "List organizations for the authenticated user",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Organization list"
          }
        }
      },
      "post": {
        "summary": "Create an organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Organization created"
          }
        }
      }
    },
    "/v1/orgs/{id}": {
      "get": {
        "summary": "Get an organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Organization"
          }
        }
      },
      "patch": {
        "summary": "Update an organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated organization"
          }
        }
      },
      "delete": {
        "summary": "Delete an organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Organization deleted"
          }
        }
      }
    },
    "/v1/orgs/{id}/members": {
      "get": {
        "summary": "List organization members",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member list"
          }
        }
      },
      "post": {
        "summary": "Add an organization member",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Member added"
          }
        }
      }
    },
    "/v1/orgs/{id}/members/{userId}": {
      "patch": {
        "summary": "Update a member's role",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Role updated"
          }
        }
      },
      "delete": {
        "summary": "Remove a member from the organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Member removed"
          }
        }
      }
    },
    "/v1/workspaces": {
      "get": {
        "summary": "List workspaces in an organization",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace list"
          }
        }
      },
      "post": {
        "summary": "Create a workspace",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "201": {
            "description": "Workspace created"
          }
        }
      }
    },
    "/v1/workspaces/{id}": {
      "get": {
        "summary": "Get a workspace",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace"
          }
        }
      },
      "patch": {
        "summary": "Update a workspace (set `archived` true/false to archive or restore)",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Updated workspace"
          }
        }
      },
      "delete": {
        "summary": "Delete a workspace",
        "tags": [
          "v1"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace deleted"
          }
        }
      }
    },
...
    "/workspaces": {
      "get": {
        "summary": "List workspaces in an organization",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "organization_id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of workspaces"
          }
        }
      },
      "post": {
        "summary": "Create a workspace",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "organization_id",
                  "title"
                ],
                "properties": {
                  "organization_id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "title": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string",
                    "description": "Optional. Generated from title if absent."
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "is_default": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Workspace created"
          }
        }
      }
    },
    "/workspaces/{id}": {
      "get": {
        "summary": "Get workspace details (with goal + plan counts)",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "patch": {
        "summary": "Update workspace title, description, icon, or slug",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "string"
                  },
                  "slug": {
                    "type": "string"
                  },
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated"
          }
        }
      },
      "delete": {
        "summary": "Hard-delete a workspace. Goals/plans inside have their workspace_id set to NULL.",
        "tags": [
          "Workspaces"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        }
      }
    },
...
  },
...
}